Float

+++ appLab4B/src/main/kotlin/com/github/hummel/cga/lab4b/Face.kt
@@ -4,5 +4,5 @@ class Face {
 	lateinit var vertices: Array<Vertex>
 	lateinit var textures: Array<Vertex>
 	lateinit var normals: Array<Vertex>
-	lateinit var depthArr: DoubleArray
+	lateinit var depthArr: FloatArray
 }
